Extract oversized inline test modules to sibling files
Extracts the oversized inline #[cfg(test)] mod tests { ... } blocks
in the three S2.3-relocated host files into sibling tests.rs files,
matching the project convention established by
crates/kernel/supervisor/src/capture.rs (and documented in AGENTS.md
§4.B: the default for out-of-line unit tests is foo.rs +
foo/tests.rs). Each of the three relocated modules carried an inline
test block well over the ~50-line threshold.

Files:

- crates/prod/core/host/src/host/buffering_invoker.rs — inline block
  (~122 lines) replaced with `#[cfg(test)] mod tests;`; body moved
  verbatim to crates/prod/core/host/src/host/buffering_invoker/tests.rs.
- crates/prod/core/host/src/host/effects.rs — inline block
  (~137 lines) replaced with `#[cfg(test)] mod tests;`; body moved
  verbatim to crates/prod/core/host/src/host/effects/tests.rs.
- crates/prod/core/host/src/host/coverage.rs — inline block
  (~114 lines) replaced with `#[cfg(test)] mod tests;`; body moved
  verbatim to crates/prod/core/host/src/host/coverage/tests.rs.

Pure layout extraction: no test logic changed; no imports changed
beyond being re-anchored at the new module root; `use super::*;`
still refers to the parent production module. Production module
shape stays as foo.rs with a sibling foo/ directory for tests,
matching capture.rs / capture/tests.rs.

These blocks were inline at the same sizes in the pre-S2.3 location
(crates/kernel/adapter/src/host/) and were carried over verbatim
during S2.3's pure-relocation scope. This commit discharges the
style debt surfaced by Greptile on PR #80.

Verification: cargo test -p ergo-host green (224 passed); cargo test
--workspace green (no failures). No §6 invocation: no §3 symbol
changes, only test-module layout.
